[Refactor] : 사용자 위도, 경도 가져오는 로직을 분리
